+++ Sources/F53OSC/F53OSCBundle.h
@@ -3,7 +3,7 @@
 //  F53OSC
 //
 //  Created by Siobhán Dougall on 1/17/11.
-//  Copyright (c) 2011-2025 Figure 53 LLC, https://figure53.com
+//  Copyright (c) 2011-2026 Figure 53 LLC, https://figure53.com
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -48,6 +48,9 @@ NS_ASSUME_NONNULL_BEGIN
 
 // redeclare as nonnull for this subclass
 - (NSData *) packetData;
+
+// deprecated
+- (nullable NSString *) asQSC DEPRECATED_MSG_ATTRIBUTE("Use -asQSC: instead.");
 
 @end