+++ Sources/F53OSC/F53OSCPacket.h
@@ -3,7 +3,7 @@
 //  F53OSC
 //
 //  Created by Siobhán Dougall on 1/17/11.
-//  Copyright (c) 2011-2025 Figure 53 LLC, https://figure53.com
+//  Copyright (c) 2011-2026 Figure 53 LLC, https://figure53.com
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -40,7 +40,10 @@ NS_ASSUME_NONNULL_BEGIN
 @property (strong, nullable) F53OSCSocket *replySocket; // If this message was received from a client, this is the socket to use to reply.
 
 - (nullable NSData *) packetData;
-- (nullable NSString *) asQSC;
+- (nullable NSString *) asQSC:(nullable NSLocale *)locale; // nil means `currentLocale`
+
+// deprecated
+- (nullable NSString *) asQSC DEPRECATED_MSG_ATTRIBUTE("Use -asQSC: instead.");
 
 @end