+++ Sources/F53OSC/F53OSCPacket.m
@@ -3,7 +3,7 @@
 //  F53OSC
 //
 //  Created by Siobhán Dougall on 1/17/11.
-//  Copyright (c) 2011-2025 Figure 53 LLC, https://figure53.com
+//  Copyright (c) 2011-2026 Figure 53 LLC, https://figure53.com
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -47,6 +47,14 @@ NS_ASSUME_NONNULL_BEGIN
     // Defined by subclasses.
     return nil;
 }
+
+- (nullable NSString *) asQSC:(nullable NSLocale *)locale
+{
+    // Defined by subclasses.
+    return nil;
+}
+
+#pragma mark - deprecated
 
 - (nullable NSString *) asQSC
 {